requirements.txt
Flask==1.1.4
Werkzeug==1.0.1
markupsafe==2.0.1
flask_restful
flask_jwt_extended
passlib
mysql.connector
email_validator
psycopg2-binary